refactor(tests): rename pems_web directory
move some pems_web specific fixtures/tests around
