+++ pyproject.toml
@@ -54,7 +54,7 @@ preserve_blank_lines = true
 use_gitignore = true
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "tests.pytest.settings"
+DJANGO_SETTINGS_MODULE = "tests.pytest.pems_web.settings"
 
 [tool.setuptools_scm]
 # intentionally left blank, but we need the section header to activate the tool
