+++ tests/pytest/conftest.py
@@ -1,17 +1,5 @@
-import pytest
 from pytest_socket import disable_socket
 
 
 def pytest_runtest_setup():
     disable_socket()
-
-
-@pytest.fixture
-def app_request(rf):
-    """
-    Fixture creates and initializes a new Django request object similar to a real application request.
-    """
-    # create a request for the path, initialize
-    app_request = rf.get("/some/arbitrary/path")
-
-    return app_request
