--- tests/pytest/pems/conftest.py
@@ -1,10 +1,0 @@
-import pytest
-
-from pems_web.districts.models import District
-
-
-@pytest.fixture
-def model_District():
-    district = District.objects.create(number="1", name="Eureka")
-
-    return district
